+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     * `validatePhoneNumber`: If `regionCode` is omitted, the phone number is validated based on the region inferred from the phone number itself
     * `parsePhoneNumber`: The `regionCode` can be safely omitted if the phone number string contains "+" followed by the country code (e.g., international, E.164, or RFC3966 formats)
     * Note that the handler needs to have already been instantiated with metadata for the expected region(s)
+  * Added `inferPhoneNumberType` and `inferPhoneNumberRegion` methods
+  * Added `inputDigitAndRememberPosition` and `getRememberedPosition` methods to the AsYouTypeFormatter
 
 ## 1.88.0